[MER-867] Add default PR title from commit subject (#27)
This also moves a lot of the code that was in `./cmd/av/pr.go` into a new folder:
`./internal/actions`. The reason for this is to be able to extract out certain
things that we'll need to do in multiple contexts (in this case, I want to be
able to implement an `av stack submit` command, which will involve opening a PR
for every branch in the stack).

That code is mostly unchanged, except for the bit where we now read the subject
and body from the commit in order to populate the PR title/body. I'm going to
submit this PR using this strategy, so... wish me luck! 🤞