Implement av stack reparent command (#25)
I originally wanted to do `av stack sync --parent <newparent>`, but it was different enough from a sync that I made it its own command. I'm still not 100% set on that but... 🤷.

Also, it's possible for there to be a rebase conflict during this. It's possible to get into an undesirable state if you then abort that rebase (because av thinks the parent is `<newparent>` but the git branch is still based on `<oldparent>`).

Nonetheless, I think I'd like to merge this as is and think about that limitation a bit later.