Fetch PR info from GH before trying to open it
This allows users to open PR's outside of `av` and then let av know about
the state of that PR. This is mostly useful in scenarios where users open
the first (root) PR using the GH UI, then realize that they want to stack
their work on top if it. Before this commit, that wasn't possible (without
closing the original PR and then re-creating it with `av pr create`).

When we implement `av stack submit`, this should also happen automatically
(for every PR in the stack), but that's a bridge we haven't quite reached.