feat(stack branch): Support renaming branches (#54)
This PR is a bit large because golangci-lint just released a new version (last night I think) that breaks our CI for some reason. I'm not entirely sure, but Go 1.19 introduced a small change to how doc comments are formatted when using `gofmt` and `golangci-lint` was complaining (in a rather opaque way) because it... is apparently trying to use that version of `gofmt` even though we're on Go 1.18? I couldn't really reproduce it with `golangci-lint`, but this change just runs the code from source directly. (Adding these kinds of things to `tools.go` is a pretty common practice and doesn't increase the size of our build executable or anything)

If you want to look at just the first commit, that might be easier since it contains all the actual code changes.