feat: Add stack metadata to PR on creation (#57)
I think this will help us be more robust against weird things happen because we're trying to figure out the users intention based on how the PRs are arranged on GitHub.

This also will mean that we don't have to use the `avbeta-stackedprs` label since we can just read the metadata from the PR comment. I think this will be necessary to avoid accidentally injecting stacked PRs functionality into existing dev workflows.

This is what it looks like:
<img width="1624" alt="image" src="https://user-images.githubusercontent.com/773453/186551027-286331f3-e587-4ddd-97ae-884aea5418fc.png">


<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track the status of stacks when using Aviator. Please do not delete or edit this section of the PR.
```
{"parent":"master","parentHead":""}
```
-->