feat: Include parent pull in stack metadata, update metadata on stack submit (#59)
This PR adds `parentPull` to stack metadata. It also will update the metadata whenever `CreatePullRequest` is called (_i.e._, when doing `av pr create` or `av stack submit`).

We'll probably need to detect if things are bad™ on the server side of things and recommend running `av stack submit`, but at least this should make it so that the logic for constructing stacks is more explicit and intentional and consistent and less likely to do unexpected things.



<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track the status of stacks when using Aviator. Please do not delete or edit this section of the PR.
```
{"parent":"master","parentHead":"","trunk":"master"}
```
-->