fix: Construct correct list of branches to be synced during stack sync (#62)
This fixes how we store the list of branches being synced during the current `av stack sync`. This bug would only manifest if the stack sync was being run from a branch that is not at the end of the stack and the stack sync was interrupted on a branch that comes after the originally checked out branch in the stack.

Confirmed that the new test failed with the code changes and passes now.



<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track the status of stacks when using Aviator. Please do not delete or edit this section of the PR.
```
{"parent":"master","parentHead":"","trunk":"master"}
```
-->