fix: Don't un-draft PRs during re-parent (#71)
Fixing this as reported by a customer.

Before this, we would actually un-draft (aka _mark ready for review_) PRs that were in a draft state if we were re-parenting the PR.

Incidentally, while fixing that bug, I realized that we actually were never transitioning PR's to a draft state at all when they should have been.

With this PR, I verified that now:
* PRs that should be switched to draft during re-parent (to avoid adding unnecessary codeowners) actually are
* PRs that are in a draft state when the re-parent runs continue to stay draft PRs