feat: Add stack branch-commit command (#76)
This change adds a new command `av stack branch-commit`. This creates a
branch and a commit in one shot. The branch name can be auto-generated
if unspecified.

The command name is long, but the user can create a git-alias for
shorthand. For example:

```
[alias]
bc = !av stack branch-commit
```

This allows the user to run `git bc -a -m "Add a new feature"`, which
ends up calling av-cli.