feat: Add av stack tidy (#85)
This command is used to update the av-cli's internal branch metadata
based on the Git branches to some extent. The expected usage is that
after a branch is merged or deleted, run this command to synchronize the
internal metadata.

Probably we can do more magic and hide the details on the branch
metadata furthermore, and automatically tidies up in av stack sync. That
can be done in the future. Right now, unable to delete a branch is
critical, and this should act as a stop-gap solution.