feat: allow editing a PR description
Previously, if there's an existing PR, an editor pops open, but the
content is discarded anyway. Now it won't opens an editor in such case.

With the new flag `--edit`, an editor opens for an existing PR. The
content is pre-filled with the existing PR description and updated
contents are reflected to the PR.