fix: Fix av CLI not re-parenting children (#107)
I think this bug was introduced in #103. I see how it happened and it suggests to me that the code is a little bit hard to follow at the moment. There's room to clean it up so that bugs like this don't reoccur. I don't think the version of `av` with this bug present has been released.

The test I added fails without this change (and I noticed this issue locally).