fix: Update the parent consistently (#111)
When a conflict happens, in some cases the parent is updated before
conflict is resolved. When sync is aborted, the parent is already moved,
so it won't get back to the original state. The parent update should be
done after --continue.

Also, the parent should be updated even if it's not trunk. It should
update the parent hash as well.

This change creates a sync continuation for all cases. The continuation
contains the data needed for updating parents. The continuation is
called inline when there's no conflict. If there's a conflict, the
continuation is saved and applied later.