feat(stack reorder): Add command scaffolding (#123)
Breaking this into a few parts.

This PR is intended to just define the base scaffolding. Namely, we're just defining the structure of the `stack-branch` and `pick` commands (ultimately the goal is to support all the commands from `git rebase --interactive` in addition to some av specific ones like `stack-branch`).

This code is not called from anywhere currently.