refactor: Remove Children from BranchMeta (#142)
Originally, we stored the children in the `meta.Branch` struct. This resulted in a duplication of the parent/child relationship (we store the parent of each branch in `Branch.Parent` as well as the children of each branch in `Branch.Children`).

This was because we had to read Git refs for each branch to get its metadata which is expensive. Now that we store the state of the repo in-memory, it's pretty trivial to just reconstruct the branch dependency graph whenever we need it.