[MER-2347] No positional args. (#153)
Has no test code.

Arguably this warrants adding a test case that verifies we error on positional args to stack sync. 

I looked at our e2e_tests for a bit and didn't have an immediate idea of where/ how to add this and I'm not yet familiar with gittest. Feel free to bounce back to me that this needs a testcase.