feat(stack reorder): Generate plan for stack (#157)
Getting pretty close to having an MVP on this feature.

The generated plan in the test case looks like
```
    plan_test.go:57: stack-branch one --trunk main@a50ac627fb10b6a7e19d39fde5f925ff9d2e9c29
    plan_test.go:57: pick 51f48a6  # Write file
    plan_test.go:57: pick baf3979  # Write file
    plan_test.go:57: stack-branch two --parent one
    plan_test.go:57: pick d856653  # Write fichier
    plan_test.go:57: pick 28fbcea  # Write fichier
```

The test cases for plan are a bit brittle at the moment, but I think we're still in the stage of figuring it out such that I think that's fine for now.