feat: Make it easy to detect a PR markdown
This is a nice to have and a harmless change. The temp file created for
the pull-request title and body now has an extension '.av.md'. This is
anyway recognized as a markdown file for most of the editors, but this
extra 'av' extension allows us to have a special file type that supports
'%%' as a comment.

As an example, added a Vim plugin that recognize this file type and
treats '%%' as a comment.