[MER-2446] Adjust styles for web and manpages
We are trying to use `docs/*` as both command manuals on the web as well
as online manual distributed along with the binary. For that, we want to
customize go-md2man output, especially the header part.

This includes a fork of https://github.com/cpuguy83/go-md2man. This
includes the title part and some whitespace handling
modifications.

Note that go-md2man is licensed under MIT license, which is the same as
this repository. The file independently have a license header and a
copyright notice with a SPDX header.