feat: Allow specifying Aviator API URL from configuration (#194)
Feature request from a customer to support on-prem installations.
Ideally this would be configurable per-repo as it's possible that a user might have some repositories that use the \~\~cloud~~ version of Aviator and some that use on-prem, but that's not very likely at this point and I didn't want to add that complexity to this pull request.