fix: Push correctly during av stack sync (#197)
Fixes a bug where branches other than the current branch in the stack would not be pushed during `av stack sync`.

The bug was essentially that `actions.Push` assumed that the branch to be pushed was the current branch and the calling code didn't match that assumption. I changed the `Push` function to take the branch name explicitly since 1) it doesn't actually require the branch to be checked out and 2) APIs with fewer non-obvious assumptions are generally better and less prone to this kind of bug.