fix: Update base branch correctly in av pr create (#206)
Fixes a bug where running `av pr create` or `av stack submit` won't edit the base branch of a pull request that already exists.

I noticed this when I reparented a branch that I'd already created a PR for and noticed it didn't set the base branch correctly on GitHub.