[MER-3739] Update libraries
Update the Golang toolchain version as well as libraries.

The workflow is broken as it's trying to install the HEAD version of
golangci-lint. This changes this to install the latest release version
instead.

In addition to this, noticed that golangci-lint hasn't been running
properly. This PR fixes that and addresses the issues pointed out by
that.