Use t.Log for test log output
The test code uses logrus / fmt.Print for output, and this makes it
difficult to debug the test failure. By using testing.T.Log, the logs
produced during the test are properly attributed to the test itself, and
we can see logs only for the failing test case.

To achieve this, the test helpers (gittest) are almost re-written to
avoid the production code for Git related operations. This also utilizes
go-git for retrieving the Git repository data, which is anyway planned
to be used in the production code as well.