Implement av stack sync v2
This is the new sequencer based av stack sync implementation.

From this version, the command does the git fetch and GitHub API call at
the beginning at once. This makes it faster than before, which runs git
fetch for every branch.

The command now asks if the user wants to push branches to remote. This
prompt is shown only when it's needed. Also this can be controlled by
the CLI flag. Also the command now asks if the user wants to delete
merged branches.

git-push will do an atomic push for all modified branches. This makes
the command faster than before, which runs git push for every branch
one-by-one.

Closes #247, #239