Add interactive `stack next` (#303)
![out](https://github.com/aviator-co/av/assets/10200358/a5bfd83e-e390-44e8-accd-905591123b25)

## Proposal

A stack next command that supports interactively selecting the path you want to navigate through. 

For example if you have multiple stacks on top of your trunk (like below) and you run `av stack next`
<img width="383" alt="image" src="https://github.com/aviator-co/av/assets/10200358/44189309-4f3d-47bc-b4b5-f7ba9f743636">


It will now prompt you to ask which one you want to use:
<img width="627" alt="image" src="https://github.com/aviator-co/av/assets/10200358/7470c6d2-bb79-47b6-90aa-97be63ea5bf1">

This is a proposal and uses the github.com/erikgeiser/promptkit library, if this is accepted and you have a preferred prompt library let me know.