Improve the stack branch command to handle trunk branches as parents
When creating a new branch from a trunk branch, the stack branch command
is better to start off from the remote tracking branch instead of the
local branch. This way, the new branch will be created from the latest
state of the trunk branch.

While we are here, accept `origin/HEAD` as the parent branch name and
resolve it to the default branch.

Previously, this command automatically "adopts" the parent branch to av
automatically, assuming that the parent branch's parent is the default
branch, which might not be true. Instead, this command checks if the
parent branch is already adopted or not.