Detect the branch structure in a more robust way
This should address the issue where the stack adopt command fails when
it's invoked at the trunk branch. (https://github.com/aviator-co/av/issues/363)

Based on the feedbacks, it seems that the commit traversal is way slower
than expected. Try to be a bit more conservative when to traverse the
commits.