Explain what stack tidy does in the output and manpages (#382)
This clarifies the `av stack tidy` behavior. It now explains which
branch was deleted and which branch was orphaned.

This alters the behavior a bit. Previously, if a parent branch is
deleted, its child branch was reparented to the grandparent. Now, the
child branches are orphaned (which sounds like an orphan, actually). The
stack tidy command explains this behavior in the output.