Use git-status to get the status of the repository (#385)(#384)
This conveniently provides the status of the repository in a structured
way. This can be used to get the current branch name (or detached
state), the OID, and all staged, unstaged, unmerged, untracked files.

Related to https://github.com/aviator-co/av/issues/374


Detect the branch structure in a more robust way
This should address the issue where the stack adopt command fails when
it's invoked at the trunk branch. (https://github.com/aviator-co/av/issues/363)

Based on the feedbacks, it seems that the commit traversal is way slower
than expected. Try to be a bit more conservative when to traverse the
commits.


Closes #384