cmd: Add av upgrade
We're testing the Aviator CLI internally and it's a bit cumbersome to manually install the package, especially in remote instances where downloading the version directly from the release page is required. This abstracts this away slightly by introducing a new  command. I've tested it on MacOS, but don't have a Windows machine / didn't setup other environments to test in. If you're interested in this, I'd recommend testing it a bit more thoroughly. I've updated the README.md and other command ergonomics such as printing out the version upgrade CTA only when not running the upgrade command itself.