Branch off from the parent branch (#529)
When creating a new branch from a trunk branch, it should branch from
the parent remote tracking branch. This hasn't been an issue if the
trunk branch is the default branch (such as main / master), but if this
is used with a additionalTrunkBranches, it branches off from main /
master always rather than the specified additional trunk branch.