Use remote tracking branch for better commit detection in adopt
This improves the accuracy of detecting the nearest trunk commit by
using the remote tracking branch. Normally, we don't directly commit
anything on local trunk branches, so its remote tracking branch is
always ahead of the local trunk branch (if it exists). This change uses
that for better commit detection when checking which commit is a part of
the topic branches.