Strip the PR metadata from the PR body
When editing an existing PR with `av pr --edit`, we fetch the existing
PR body from GitHub and launch an editor to edit it. However, the PR
body on GitHub might already has a PR metadata section / PR stack info
appended to it, which we don't want to show in the editor. Strip them
from the body.