Strip the PR metadata from the PR body (#598)(#597)(#595)
When editing an existing PR with `av pr --edit`, we fetch the existing
PR body from GitHub and launch an editor to edit it. However, the PR
body on GitHub might already has a PR metadata section / PR stack info
appended to it, which we don't want to show in the editor. Strip them
from the body.



Use remote tracking branch for better commit detection in adopt
This improves the accuracy of detecting the nearest trunk commit by
using the remote tracking branch. Normally, we don't directly commit
anything on local trunk branches, so its remote tracking branch is
always ahead of the local trunk branch (if it exists). This change uses
that for better commit detection when checking which commit is a part of
the topic branches.



Closes #597

Create a convenience function for handling errors




Closes #595