Remove golines and add gofupt to pre-commit config (#596)
golines is unmaintained and is going to be archived in Q4 2025
(https://github.com/segmentio/golines). Long lines are not a big issue
in Go code, and this occasionally introduces an odd formatting style.
Remove this.

While we are here, add gofumpt, which is a stricter version of gofmt. We
already use it as a part of golangci-lint, so this will ensure that the
format is checked before committing rather than later in CI.