Run pre-av-sync hook with tty (#601)
Previously the hook was executed while av having a control on tty. This
means that the hook cannot be interactive. This commit changes the hook
execution to be done with bubbletea's ExecProcess, which allows the
hook to take over the underlying tty. This allows users to create
interactive hook, which was not possible before.

There was a try-and-error on the visual of the hook execution. Details
are written in the comment of preAvSyncHookModel.View(). This will not
change anything if there is no pre-av-sync hook. If there is a hook, it
will show a pre-av-sync hook status shown in the sync view.