fix: editor var expansion (#603)
This handles cases where the git editor is set to `$EDITOR`. The steps for reproducing are as follows.

1. Set the EDITOR env var
2. Update your .gitconfig to include editor=$EDITOR
3. Run `av pr --edit`

According to `man 1 git-var`, `GIT_EDITOR` is "meant to be interpreted by the shell when it is used". Hence, this variable needs to be interpreted.

Credit to @tgandrews for spotting the issue