Open editor when there's no title and no body (#607)
Previously, we didn't strip the PR metadata from the existing PR body,
but with recent changes, we do.

As a side effect, if an existing PR doesn't have a body (but has a
title), it would open the editor unnecessarily with `av pr` command.
Open it only when both title and body are empty.