Use continuation instead of message to signal subcomponent results (#604)
Previously, `av sync` was using message passing to signal subcomponent
results. By sending out a special message to the bubbletea message bus,
the parent component can tell if a subcomponent finishes its process.
For example, sequencerui had three messages, Conflict, Abort, and Done,
to indicate its processing result.

While this works, it becomes complicated to handle those messages. As we
add more subcomponents, we need to handle the messages in a central
message handling function.

Instead of using message passing for signaling the results, use
continuation passing style to handle the results. This style is already
employed in the official API such as ExecProcess
(https://pkg.go.dev/github.com/charmbracelet/bubbletea#ExecProcess).

The benefit of the continuation passing style in `av` is that we can
write the processing flow from top to bottom. The passed continuation
indicates what will be done after one step. This reduces the number of
messages that we need to handle in the message handler.

Also, because `av` is doing processing step-by-step, it turned out that
we just can forward most of the messages to the last subcomponent. This
greatly simplifies the message handling code as well as the rendering
code.

Overall, this reduces the complexity of the implementation structure.