Support specifying a branching point even if the parent is trunk (#623)(#622)(#621)(#620)(#619)(#618)
This allows us to handle a case where a newly adopted remote branch has
an already merged branch as a parent. In the next PR, we will specify
the branching point for those branches.


Check if the default branch is set when opening a repository
We can check if the default branch at remote tracking branch
(refs/remotes/origin/HEAD) exists when opening a repository. By doing
so, we don't have to take context and do not have to do an error
handling in many situations.


Closes #622

Allow specifying the branching point even for trunk parents
This relax the semantics of the branch metadata. In the branch metadata,
we track `trunk: bool` and `head: string`, but only one of them could be
specified in the current semantics. This update allows them to specify
both. That is, even if the parent branch is a trunk branch, a branch
metadata can specify the branching point. Accordingly, we update the
internal name of that branching point as BranchingPointCommitHash.

Newly added av branch --remote motivates this change. In that command,
we might see a partially merged stack. In that case, we will have a case
where a topic branch would have a trunk branch as a parent (because its
original parent has been merged), but at the same time, such topic
branches would need to have an explicit branching point in order to
avoid already merged commits to be considered as a part of that topic
branch.

Behavior updates would be necessary in the sequencer, but we will do
that in the following PRs.


Closes #621

Deprecate reference based metadata store
It's been a while since we migrated to a file-base metadata storage. We
can drop this refmeta code.



Closes #620

Extract MergeCommit for remote branches




Closes #619

Handle a case where no branch was chosen




Closes #618