Handle a case where a stack is partially merged (#624)
There's a case where a stack is partially merged and not `av sync`ed. In
this case, the parent branch of the not-yet-merged branch no longer
exists on GitHub in many cases (but we can get the PR info via GraphQL
API). The current implementation fails on git-fetch as the branch
doesn't exist.

This change skips those merged branches and adopt only branches that are
still open. The branch metadata is (locally) updated to have a trunk
branch as the parent.

There was an issue that we couldn't restack those branches, but with a
series of changes around the branching point (such as
https://github.com/aviator-co/av/commit/e3a22062a5734108a7ddfb3345735d1c3b750168)
makes it possible to support this case.