Allow adopting branches on top of local branches (#625)
There's a case where a branch already has been adopted locally, and they
want to adopt a new branch on top of that local branch. This change
allows that action.

While we are here, after a successful adoption, we check out the leaf
branch so that the user is in a good state to continue working.